Clean up the code base

+++ src/scripts/inference_pretained.py
@@ -14,7 +14,7 @@ from data.data_loader import LotkaVolterraDataset, encode, decode
 def main():
     model_name = "Qwen/Qwen2.5-0.5B-Instruct"
     tokenizer = AutoTokenizer.from_pretrained(model_name)
-    model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype=torch.float16, device_map="cpu")
+    model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype=torch.float32, device_map="cpu")
     model.eval()
